requirements.txt
# Python 3.11.9
fastapi==0.111.0
uvicorn==0.29.0
pandasai==1.5.0
pandasai-litellm==0.1.16
python-dotenv==1.0.1